2-Tabla-Reservas-Admin

+++ ClubDeportivo/src/main/java/com/salesianos/triana/dam/clubDeportivo/repository/SocioRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.salesianos.triana.dam.clubDeportivo.model.Socio;
 
-public interface SocioRepositorio extends JpaRepository <Socio, Integer>{
+public interface SocioRepositorio extends JpaRepository <Socio, Long>{
 
 }
