+++ ClubDeportivo/src/main/java/com/salesianos/triana/dam/clubDeportivo/service/SocioService.java
@@ -6,6 +6,6 @@ import com.salesianos.triana.dam.clubDeportivo.model.Socio;
 import com.salesianos.triana.dam.clubDeportivo.repository.SocioRepositorio;
 
 @Service
-public class SocioService extends BaseServiceImp<Socio, Integer, SocioRepositorio>{
+public class SocioService extends BaseServiceImp<Socio, Long, SocioRepositorio>{
 
 }
